+++ CHANGELOG.md
@@ -6,6 +6,8 @@ There are some key used within the changelog which are as following:
 
 ## IN PROGRESS
 
+- added badge components #126
+- added breadcrumbs components #117
 - added notification components #129
 - added progress bar components #132
 - added modal components #128